fix(users | todos): Убрал лишние сервисы и фасад
