+++ src/app/libs/users/pages/users/users.component.ts
@@ -2,10 +2,10 @@ import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { User } from '../../interfaces/users.interface';
 import { UsersListComponent } from '../../components/users-list/users-list.component';
 import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
-import { UsersFacade } from '../../users.facade';
 import { Observable } from 'rxjs';
 import { AsyncPipe } from '@angular/common';
 import { UserCreateFormComponent } from '../../forms/user-create-form/user-create-form.component';
+import { UsersService } from '../../services/users.service';
 
 @Component({
   selector: 'app-users',
@@ -16,27 +16,32 @@ import { UserCreateFormComponent } from '../../forms/user-create-form/user-creat
   changeDetection: ChangeDetectionStrategy.OnPush,
 })
 export class UsersComponent {
-  private readonly usersFacade: UsersFacade = inject(UsersFacade);
+  private readonly usersService: UsersService = inject(UsersService);
 
-  readonly users$: Observable<User[]> = this.usersFacade.users$;
+  readonly users$: Observable<User[]> = this.usersService.users$;
 
   constructor() {
     this.initUsers();
   }
 
   createUser(newUser: User): void {
-    this.usersFacade.createUser(newUser);
+    if (this.usersService.isExistingUserEmail(newUser)) {
+      alert('Такой email уже существует!');
+    } else {
+      this.usersService.createUser(newUser);
+      alert('Новый пользователь успешно создан!');
+    }
   }
 
   editUser(editedUser: User): void {
-    this.usersFacade.editUser(editedUser);
+    this.usersService.editUser(editedUser);
   }
 
   deleteUser(userId: number): void {
-    this.usersFacade.deleteUser(userId);
+    this.usersService.deleteUser(userId);
   }
 
   private initUsers(): void {
-    this.usersFacade.initUsers().pipe(takeUntilDestroyed()).subscribe();
+    this.usersService.initUsers().pipe(takeUntilDestroyed()).subscribe();
   }
 }
